backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# LangGraph and LangChain dependencies
langgraph==0.0.20
langchain==0.1.0
langchain-openai==0.0.2
langchain-google-genai==0.0.6

# LLM providers
openai==1.3.7
google-generativeai==0.3.2

# Environment and utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Database (for future use)
supabase==2.0.2

# Logging and monitoring
structlog==23.2.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
